Update Chart.swift
XCode Beta 3 started giving me a "A non-failable initializer cannot chain to failable initializer 'init(coder:)' written with 'init?'" as an error on this line. Changing Init to Init? as Xcode recommends will get Xcode to stop throwing an error. However, I assume that we still need to handle the nil case.

Here is a related question on stackoverflow:

http://stackoverflow.com/questions/26427007/failable-initializer-in-xcode-6-1

 Can anyone confirm that this is infact new with beta 3 or is it just a symptom of a problem somewhere else in my code?