I am adding an Optional completion handler for TouchHighlight fade animation.
This was submitted previously, however there were issues when squashing. This was the previous description:
Having an optional completion handler for the TouchHighlight fade animation should come in handy. For example, if there are elements that change when a user starts panning, this will allow us to revert back to the original state. For example, you can have two fixed UILabels at the top of the view that correspond to a charts below it. They are initially set to the last X and Y values of the chart. As the user pans these values change. With this completion handler, the UILabels can switch back to the original values when user stops the pan (after the fade animation). This is just one example but I'm sure it can be useful. Let me know if you would like to change the naming convention.

Changed naming convention and other minor formatting/convention improvements.